+++ chef-repo/site-cookbooks/develop/attributes/firewalld.rb
@@ -1,4 +1,0 @@
-include_attribute 'firewalld::default'
-
-default['firewall']['services'] = ["ssh", "ntp", "http", "https"]
-default['firewall']['ports'] = []