+++ chef-repo/site-cookbooks/develop/attributes/laravel.rb
@@ -1,2 +1,0 @@
-default['laravel']['servername'] = "laravel.local"
-default['laravel']['path'] = "/home/vagrant/laravel"