+++ chef-repo/site-cookbooks/develop/attributes/rails.rb
@@ -1,3 +1,0 @@
-default['rails']['servername'] = "rails.booking.com"
-default['rails']['path'] = "/home/vagrant/rails"
-default['rails']['railsenv'] = "development"