+++ chef-repo/site-cookbooks/develop/attributes/yum-ius.rb
@@ -1,3 +1,0 @@
-include_attribute 'yum-ius::ius'
-
-default['yum']['ius']['enabled'] = false