+++ chef-repo/site-cookbooks/devtools/metadata.rb
@@ -1,15 +1,9 @@
-name             'develop'
+name             'devtools'
 maintainer       'YOUR_COMPANY_NAME'
 maintainer_email 'YOUR_EMAIL'
 license          'All rights reserved'
-description      'Installs/Configures develop'
+description      'Installs/Configures devtools'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
-depends 'yum'
-depends 'yum-remi'
-depends 'yum-epel'
-depends 'ntp'
 depends 'firewalld'
-depends 'nginx'
-depends 'poise-python'