+++ chef-repo/site-cookbooks/devtools/CHANGELOG.md
@@ -1,9 +1,9 @@
-# dev CHANGELOG
+# devtools CHANGELOG
 
-This file is used to list changes made in each version of the dev cookbook.
+This file is used to list changes made in each version of the devtools cookbook.
 
 ## 0.1.0
-- [your_name] - Initial release of dev
+- [your_name] - Initial release of devtools
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.